Nicolas Regnault: - yet another series of bug fixes
git-svn-id: https://www.nick-ux.org/diagham/svn/trunk@3615 b9410c57-6ef9-0310-98bc-a73ba792996a

+++ src/Vector/ComplexVector.cc
@@ -1787,7 +1787,7 @@ ComplexVector& ComplexVector::Multiply (const HermitianMatrix&  M, ComplexVector
     {
       this->Components[i].Re = M.DiagonalElements[i] * V.Components[i].Re;
       this->Components[i].Im = M.DiagonalElements[i] * V.Components[i].Im;
-      int pos = i - 1;
+      long pos = i - 1;
       int j = 0;
       for (; j < i; ++j)
 	{
@@ -1795,7 +1795,7 @@ ComplexVector& ComplexVector::Multiply (const HermitianMatrix&  M, ComplexVector
 	    M.ImaginaryOffDiagonalElements[pos ] * V.Components[j].Im;
 	  this->Components[i].Im += M.RealOffDiagonalElements[pos] * V.Components[j].Im 
 	    - M.ImaginaryOffDiagonalElements[pos] * V.Components[j].Re;
-	  pos += this->Dimension - j - 2 + M.Increment;
+	  pos += (long) (this->Dimension - j - 2 + M.Increment);
 	}
       ++pos;
       ++j;
@@ -1831,7 +1831,7 @@ ComplexVector& ComplexVector::Multiply (const HermitianMatrix&  M, ComplexVector
   double x;
   double y;
   int i = 0;
-  int Pos = sourceStart - 1;
+  long Pos = sourceStart - 1;
   for (; i < sourceStart; ++i)
     {
       x = 0.0;
@@ -1844,13 +1844,13 @@ ComplexVector& ComplexVector::Multiply (const HermitianMatrix&  M, ComplexVector
 	    + M.ImaginaryOffDiagonalElements[Pos] * V.Components[j].Re;
 	  ++Pos;
 	}
-      Pos += Inc1 - i;
+      Pos += (long) (Inc1 - i);
       this->Components[i].Re = x;
       this->Components[i].Im = y;
     }
   Inc1 = this->Dimension - Last + M.Increment;
-  int Pos2 = Pos;
-  int Pos3 = Pos;
+  long Pos2 = Pos;
+  long Pos3 = Pos;
   ++Pos2;
   for (; i < Last; ++i)
     {
@@ -1863,7 +1863,7 @@ ComplexVector& ComplexVector::Multiply (const HermitianMatrix&  M, ComplexVector
 	    M.ImaginaryOffDiagonalElements[Pos] * V.Components[j].Im;
 	  y += M.RealOffDiagonalElements[Pos] * V.Components[j].Im 
 	    - M.ImaginaryOffDiagonalElements[Pos] * V.Components[j].Re;
-	  Pos += Inc2 - j;
+	  Pos += (long) (Inc2 - j);
 	}
       x += M.DiagonalElements[i] * V.Components[i].Re;
       y += M.DiagonalElements[i] * V.Components[i].Im;
@@ -1876,7 +1876,7 @@ ComplexVector& ComplexVector::Multiply (const HermitianMatrix&  M, ComplexVector
 	    M.ImaginaryOffDiagonalElements[Pos2] * V.Components[j].Re;
 	  ++Pos2;
 	}
-      Pos2 += Inc1; 
+      Pos2 += (long) Inc1; 
       ++Pos3;
       this->Components[i].Re = x;
       this->Components[i].Im = y;
@@ -1892,7 +1892,7 @@ ComplexVector& ComplexVector::Multiply (const HermitianMatrix&  M, ComplexVector
 	    M.ImaginaryOffDiagonalElements[Pos] * V.Components[j].Im;
 	  y += M.RealOffDiagonalElements[Pos] * V.Components[j].Im - 
 	    M.ImaginaryOffDiagonalElements[Pos] * V.Components[j].Re;
-	  Pos += Inc2 - j;
+	  Pos += (long) (Inc2 - j);
 	}
       ++Pos3;
       this->Components[i].Re = x;
